Keyparser added
This classifies user-inputs as NEXT, PREVIOUS, EXIT, or ENTER. With this, I can make the logic for tracking slide numbers based on the order of these actions. It outputs both a full keylog txt file and an action keylog txt file. The action keylog will be read by my next script that will determine the timestamps of the slides.